// apps/api/src/service-pins/dto/update-service-pin.dto.d.ts
export declare class UpdateServicePinDto {
    positionType?: string;
    count?: number;
    shiftType?: string;
    startTime?: string;
    endTime?: string;
    days?: string;
    payRate?: number;
    specialInstructions?: string;
    geoLat?: number;
    geoLng?: number;
    status?: string;
}